+++ release_notes.md
@@ -6,6 +6,7 @@
 -   APP-5037 - [API] Updated V3 API to include new endpoints and/or updates to existing endpoints
 -   APP-5055 - [Services] Updated Paho MQTT library and message broker reconnect logic
 -   APP-5067 - [Services] Fix issue with redis connection on each new thread
+-   APP-5068 - [Batch] Added auto truncate of attribute values
 
 ## 4.0.10