+++ release_notes.md
@@ -7,6 +7,7 @@
 -   APP-5055 - [Services] Updated Paho MQTT library and message broker reconnect logic
 -   APP-5067 - [Services] Fix issue with redis connection on each new thread
 -   APP-5068 - [Batch] Added auto truncate of attribute values
+-   APP-5072 - [Services] Applied PR #356
 
 ## 4.0.10